Adding new steps to build and deploy the Sail Operator and deploy also an Istio control plane

Fix error in script


Adding skip for maistra pod creation if maistra builder image is empty


Adding debug information


Fix path for build and push script


Adding debug information


Adding sail opertor installation from nightly channel


Adding sleep before waiting for the operator


Adding a retry logic to wait for the operator


Update from rebase


Update exit code after deploying the operator


Fix variable name

+++ ci-operator/step-registry/servicemesh/istio/post/servicemesh-istio-post-commands.sh
@@ -4,5 +4,13 @@ set -o nounset
 set -o errexit
 set -o pipefail
 
-oc delete project "${MAISTRA_NAMESPACE}"
-echo "Deleted \"$MAISTRA_NAMESPACE\" Namespace"
+# Check if MAISTRA_NAMESPACE exists before attempting to delete it
+if oc get namespace "${MAISTRA_NAMESPACE}" > /dev/null 2>&1; then
+    echo "Namespace \"${MAISTRA_NAMESPACE}\" exists. Proceeding to delete it."
+    oc delete project "${MAISTRA_NAMESPACE}"
+    echo "Deleted \"$MAISTRA_NAMESPACE\" Namespace"
+else
+    # Skip deletion if the namespace does not exist because it means that servicemesh-istio-e2e step was skipped
+    echo "Namespace \"${MAISTRA_NAMESPACE}\" does not exist. No need to delete."
+    exit 0
+fi
